+++ app/models/tasks.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, Date, DateTime, ForeignKey, String, Text, UniqueConstraint
+from sqlalchemy import Column, Date, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.sql import func
 
 from app.database import Base
@@ -7,10 +7,10 @@ from app.database import Base
 class Task(Base):
     __tablename__ = "tasks"
 
-    id = Column(BigInteger, primary_key=True, index=True)
-    project_id = Column(BigInteger, ForeignKey("projects.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, index=True)
+    project_id = Column(Integer, ForeignKey("projects.id", ondelete="CASCADE"), nullable=False)
     # 担当者は未アサイン(NULL)を許容し、ユーザー削除時は担当者のみNULLにする(SET NULL)
-    assigned_user_id = Column(BigInteger, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
+    assigned_user_id = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
     title = Column(String(128), nullable=False)
     status = Column(String(20), server_default="todo", nullable=False)
     start_date = Column(Date, nullable=True)
@@ -24,9 +24,9 @@ class Task(Base):
 class Comment(Base):
     __tablename__ = "comments"
 
-    id = Column(BigInteger, primary_key=True, index=True)
-    task_id = Column(BigInteger, ForeignKey("tasks.id", ondelete="CASCADE"), nullable=False)
-    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, index=True)
+    task_id = Column(Integer, ForeignKey("tasks.id", ondelete="CASCADE"), nullable=False)
+    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     content = Column(Text, nullable=False)
 
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
@@ -36,7 +36,7 @@ class Comment(Base):
 class Label(Base):
     __tablename__ = "labels"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     name = Column(String(50), unique=True, nullable=False)
     color_code = Column(String(7), nullable=True)
 
@@ -47,9 +47,9 @@ class Label(Base):
 class TaskLabel(Base):
     __tablename__ = "task_labels"
 
-    id = Column(BigInteger, primary_key=True, index=True)
-    task_id = Column(BigInteger, ForeignKey("tasks.id", ondelete="CASCADE"), nullable=False)
-    label_id = Column(BigInteger, ForeignKey("labels.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, index=True)
+    task_id = Column(Integer, ForeignKey("tasks.id", ondelete="CASCADE"), nullable=False)
+    label_id = Column(Integer, ForeignKey("labels.id", ondelete="CASCADE"), nullable=False)
 
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now(), nullable=False)
